recall/recall-api/requirements.txt
numpy<2.0
fastapi==0.111.0
uvicorn[standard]==0.30.0
faiss-cpu==1.11.0     
redis>=4.3.4
pyarrow==14.0.1
pandas

